+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [v0.0.3] - 2025-01-19
+### What's New 🎉~
+- Feature 1: SanitizeToCamelCase function added to be used for CSS ids etc
+- _Some_ tests added
+
 ## [v0.0.2] - 2025-01-18
 ### What's New 🎉~
 - Feature 1: File, Math and Text utilities moved to their own packages